Update keyboard.c
This version of Keyboard.c has a modularized version of the execute keystrokes mechanics. Where the calling of the CALLBACK_HID_Device_CreateHIDReport() function is done from a separate function that puts the desired keystrokes into the scanCodes Array and calls the needed function as many times as needed sequentially.